+++ packages/app/src/features/dialogs/migrate/upgrade/e2e/DaiToNst.test-e2e.ts
@@ -3,8 +3,8 @@ import { SavingsPageObject } from '@/pages/Savings.PageObject'
 import { setupFork } from '@/test/e2e/forking/setupFork'
 import { setup } from '@/test/e2e/setup'
 import { test } from '@playwright/test'
-import { DialogPageObject } from '../../common/Dialog.PageObject'
-import { UpgradeDialogPageObject } from './UpgradeDialog.PageObject'
+import { DialogPageObject } from '../../../common/Dialog.PageObject'
+import { UpgradeDialogPageObject } from '../UpgradeDialog.PageObject'
 
 test.describe('Upgrade DAI to NST', () => {
   const fork = setupFork({ chainId: NST_DEV_CHAIN_ID, simulationDateOverride: new Date('2024-08-05T10:43:19Z') })
